[FIX] sale_subscription: prevent change of `recurring_invoice` field
Issue:
------
It is possible to modify the `recurring_invoice` field
even if the product has already been sold.
After closing the error message, it is possible to save
the product with the modifications.

Cause:
------
Even if we return an error message (>< raising an exception),
the value is changed in the backend.

opw-3483040

closes odoo/enterprise#46571

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>