[FIX] account_disallowed_expenses: fix category code constraint
This commit is a complement to this PR : https://github.com/odoo/enterprise/pull/46505

We are simply fixing the constraint that was never triggered since `company_id` is not required and has no default value.
So we decided to make the constraint global instead of company dependent.

opw-3463858

closes odoo/enterprise#46584

Related: odoo/upgrade#5092
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>