[FIX] hr_payroll: worked hours timezone issue
Steps to reproduce:

- Install Attendances app and Payroll app
- Setup an employee with a contract which work entry source is from attendance
- Change the working schedule to different timezone, for example Asia/Hong_Kong
- By default the working schedule should be start to work from 0800 to 1700
- Now create an attendance in the first day of the contract, i.e. 1st May 2023,
check in time 0700, check out time 2000
- The total work hours should be 13 hours
- Now create a payslip for this employee
- Check the worked days total number of hours is 12 hours

Current behaviour:
The worked days total number of hours mismatches the number of work hours in
attendance

Expected behaviour:
The worked days total number of hours should be same as the number of work hours
in attendance

Explanation:
This issue is caused by the timezone issue. When the system try to call
_get_work_hours() method, it will calculate a date_from datetime with
datetime.min.time(). However this datetime didn't consider the timezone.
Therefore it stays in UTC timezone. In the meantime, the work_entry is in, i.e.
Asia/Hong Kong. Plus the system will compare the start time of the work entry and
the date_from datetime to get the larger one. Therefore, the worked hours calculation
is wrong and the value mismatched the attendance.

task-3468012

closes odoo/enterprise#46574

Related: odoo/odoo#133610
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Yannick Tivisse <yti@odoo.com>