[IMP] account_reports: add date options for tax periods
Currently the tax report always opens the previous month by default.
Since tax periods are not always months, it makes much more sense to
open the previous "tax period" (whatever length it has) when opening the
tax report.

In order to allow that, this commit introduces the calculation and
selection of tax periods on reports. The rest of the functionality is
done in the related community commit.

task-3481913

closes odoo/enterprise#46601

Related: odoo/odoo#133715
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>