[FIX] l10n_de_reports: tax report updated to 2023
The format changed in 2023.
Most values were already changed, but the version was not.
But we also put the values of the company in the right field now.
We don't remove them from the 'DatenLieferant', as some people
could rely on it, and it does nothing when exporting it by hand.

task-3486730

closes odoo/enterprise#47181

X-original-commit: 79b3b65f1d30f081ace3c3d591b5ad4dd34fb917
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>