[FIX] l10n_de_reports: tax report updated to 2023
The format changed in 2023.
Most values were already changed, but the version was not.
But we also put the values of the company in the right field now.
We don't remove them from the 'DatenLieferant', as some people
could rely on it, and it does nothing when exporting it by hand.

task-3486730

closes odoo/enterprise#47271

X-original-commit: 4c7956c128576188040fa7aad293c070f5e35bd3
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>