[FIX] knowledge: allow sorting under readonly parent
Allow a user with read access on a parent and write access on its children to
reorder the children in the Knowledge sidebar.

The `write` method of the knowledge article model is adjusted to not throw an
access error if it receives a `parent_id` val which is the same as the current
`parent_id` as the current article, even if the current user does not have write
access on the parent.

Change the closing string for sidebar errors to "Close" and translate the title,
and update the `.pot` file

task-3456741

closes odoo/enterprise#47640

Related: odoo/odoo#135981
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>