[IMP] account_reports: avoid breaking exports with groupby customisations
From 16.0 on, it is possible for the user to change the groupby value of every report line, giving a lot of flexibility to the user in the way the reported data are displayed. The problem is, sometimes, dedicated file exports exist for these report, and depend on the value of this groupby field, considering it always has the value defined in the data files. In such cases, customizing the groupby of some lines will break the export.

We solve that by introducing a new user_groupby field, duplicating the content of the groupby at report creation, and manually editable by the user. The print_mode option also disappears in favor of an 'export_mode' key, telling, when we're exporting, what the source of the export is. When the export_mode tells us we're generating a file, we know we must use groupby instead of user_groupby generate the content of the report.

Task 3413437

Part-of: odoo/enterprise#46657