[IMP] account_reports: add info banner on top of form view for reports using a custom handler
When customizing reports by hand from the UI, one need to be careful not breaking some behavior enforced by a custom handler. To raise better awareness of this, we add a banner on top of the reports' form view informing the user of the presence of a custom handler.

Task 3413437

closes odoo/enterprise#46657

Related: odoo/odoo#133830
Signed-off-by: John Laterre (jol) <jol@odoo.com>