[IMP] planning: display the progress bar for all project
This commit's purpose is to fix the display of progress bar in the gantt
view.

before this commit:
When the planning view is grouped by project, the progress bar is not
correctly displayed for some project

after this commit:
The progress bar is now correctly displayed

step to reproduce:
- go to the project setting
- enable the planning option
- open the planning app
- click on the 'schedule by project' menu
The progress bar of project with allocated hours but without any slot on
the date range of the view is not displayed.

solution:
update the data send to the js rpc call of the gantt_progress_bar
method.

detail of the implementation:
The dict send to the javascript side is created in the method
gantt_progress__bar_project_id of the 'planning.slot' moduel. It used to
only use the result of the read_group of the 'planning.slot' to create the
return values. It was updated to also take into account the project that
were not in the read_group result.

task-3485773

closes odoo/enterprise#49403

X-original-commit: 482ea0493115694e33d148a756ef497564e10539
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>