[IMP] account_accountant: Bank Reco - Visualize selected line
Problem
---------
In the Bank Reco Widget, we visualize selected lines when relevant to
the interface. For example, the selected transaction is in blue on the
right. In the Matching tab, if a line is selected, it's in blue as well.
But when you use the "Manual Operations tab", there is no such
information. Despite the fact you can click on the different lines above
and visualize each of them in the "Manual Operations" tab (and they have
different values / fields), the line above is not highlighted.

Objective
---------
When the tab "Manual Operations" is opened, highlight the selected line
above to improve overall clarity.

Solution
---------
1. In the SCSS file, modify the background color of selected line from
the tag `o_bank_rec_selected_line`.
2. Add a condition so that the tag is not given to the selected line
when the notebook page is not the Manual operation tab one.

task-3482064

closes odoo/enterprise#47730

X-original-commit: 38bff842db3cf961f89e38baf1e42828f26d6afc
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>