[IMP] spreadsheet_{edition,dashboard}: remove locale RPC
When the user locale is different from the spreadsheet locale,
a small the globe warning icon appears.

Currently, the user locale is fetched with an RPC everytime
the user opens a spreadsheet. RPC can introduce a significant
latency to open the spreadsheet.

With this commit, the user locale is fetched with all the other
spreadsheet data. No need for an extra RPC.

closes odoo/enterprise#46460

Task: 3483938
Related: odoo/odoo#133500
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>