[REF] {document_}spreadsheet_{edition,dashboard}: group common props
The abstract spreadsheet edition is implemented 3 times by 3 actions (regular
spreadsheet documents, templates and dashboards).
The template is redefined in each of these 3 actions to account for their
specificities.
The generic props for the control panel must be given in each of these
templates. If one props is added, all 3 templates must be updated.

This commit groups all props for the control panel in the abstract action.
They can be used by `t-props="controlPanelProps"` without actually knowing
the details about generic props.

Part-of: odoo/enterprise#46460