[FIX] sale_subscription: action_update_prices did not computed prices properly
Before this commit, when the action_update_prices action was called, all the prices
were not properly computed if the line already had a parent_line_id.
As calling this action should recompute completely the prices regarless of
parent prices, we force the price to 0 first to break link to parent_line_id
and we let the normal logic determine the price of the line.

closes odoo/enterprise#46729

X-original-commit: 8bfd432ef3bbb239b6a1edcabe4bebbab3a7f35d
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>