[IMP] account_accountant: Hide match button in journal items to reconcile
Problem
---------
With the new Manual Reconciliation update the Match button just applies
a filter. Once the filter is applied you can hit match again and it will
just apply the same filter again and essentially do nothing. Seems
confusing for customers.

Objective
---------
When you are in the context of "Journal Items to Reconcile" -> don't
show "Match" in the column Matching. We need to keep the column if you
go to a matching number existing.

Solution
---------
Create a JS function that returns true if both `partner_id` and
`account_id` are used as filters at the same time. If true, the matching
button is shown.

task-3488287

closes odoo/enterprise#46654

Signed-off-by: Quentin De Paoli <qdp@odoo.com>