[REF] spreadsheet_edition: use confirmation dialog for errors
When a error dialog was raised by the spreadsheet, the custom
`SpreadsheetDialog` was used even though it's doing the same thing as the
generic `ConfirmationDialog` component in this case (given its props).

Using `ConfirmationDialog` here allows to simplify the `SpreadsheetDialog`.
It will only need to handle the specific need.

Note: the `callback` argument is a dead argument and is removed.
Part-of: odoo/enterprise#46697