[FIX] documents: prevent deleting shares when sharing
Since odoo/odoo#128244, when in the workspace share popup

1. A user clicks on `Share`
2. `onClose` (via `saveRecord`) is called before our `onSave` handler
3. the flag `saved` is `false` and the share record is **deleted**
4. Bonus: the user gets a link thinks all is going fine,
  but of course it leads to nowhere.

Note that onSavedRecords works anyway to solve this, so we can clean
our class extension, but not completely as we're in stable.

We take this opportunity to rename the cancel button into "Discard",
so users can figure out that using the "Copy Link" button does not
mean that the share will exist. I.e. if one clicks on "Copy link",
then "Cancel" or "X", they also have a link to a deleted record.

Task-3489334

closes odoo/enterprise#46690

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>