[FIX] l10n_be_reports: empty lines
During this commit: https://github.com/odoo/enterprise/commit/2d774344109c2f9b55d9ea6160ae99aeca4f0420
We've made sure that a line at level 0 has an empty line (empty tr) above it to
make the reports easier to read.

In the case of Belgian profit and loss, the empty lines were already hardcoded
in xml, which conflicted with the modification explained above. In this commit,
we remove the various empty lines.

Also, we removed the hardcoded total lines in the report and places the
result of these lines in the "ASSETS" and "EQUITY AND LIABILITIES" lines.

Task: 3437257
Part-of: odoo/enterprise#45792