[REF] spreadsheet_edition: simplify spreadsheet dialog
Component `SpreadsheetDialog` is doing multiple things:
- it acts as a dialog with an input (with props `edit: true`)
- it acts very similarly to a confirmation dialog (with props `edit: false`)

Now that every uses of the component with `edit: false` have been replaced
with the confirmation dialog (see previous commits), `SpreadsheetDialog`
can be simplified to do only one thing: being a dialog with an input.

props `errorText` is removed since it was unused.

The component is renamed to better reflect what it does. The
files are renamed accordingly in the next commit to keep the
files history.

Part-of: odoo/enterprise#46697