[FIX] l10n_fr_fec_import: handle `Idevise` with missing `Montantdevise`
Summary
-------
Fixes an issue where importing a FEC file containing lines with a
defined `Idevise` but missing `Montantdevise` results in journal items
with zero debit and credit.

Setup
-----
* install `l10n_fr_fec_import`
* switch to a French company

Steps to reproduce
------------------
* Import a FEC file with lines as described above.
* go to the journal items and find the imported lines

You should see that they incorrectly show a debit and credit of 0.

Cause
-----
When `Idevise` is defined but `Montantdevise` is missing, the system
assigns to the line an `amount_currency` value of `0.0`. This triggers
the `_inverse_amount_currency` method, which sets the balance, and
subsequently the debit and credit, to zero.

opw-3383380

closes odoo/enterprise#47090

X-original-commit: b0b5fd0a7039a6462682062c4f2fd959fc4f9ae0
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>