[FIX] mrp_workorder: displayInstruction to correct step
Currently the display instruction was base on current_check_id field.
However this field was use in the old tablet view as a hack for
navigation but now we don't need it anymore. So well just base the
next instruction on checks order with next_check_id and find if a
step hasn't be pass yet

Part-of: odoo/enterprise#46701