[FIX] point_of_sale: total due converted in pos currency
Current behavior:
If a PoS had a different currency than the company currency, the total
due was not converted in the PoS currency. So when you paid the total
due it was not the correct amount.

Fix:
The total due is now converted in the PoS currency when the PoS is
loading the data. We also had to change the refreshTotalDueOfPartner to
retrieve the total due in the PoS currency.

Steps to reproduce:
- Create a PoS that uses a different currency than the company currency
  (change the journal, pricelist and payment method)
- Make some payments with this PoS using the customer account
- Check the total due of the customer account in the PoS, it's not the
  correct value.
- Try to pay the total due, it's not the correct value either.

opw-3469682

closes odoo/enterprise#47851

X-original-commit: 2fe9393f9aaef84018fb60cfa3a5868982151696
Related: odoo/odoo#136528
Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>