[FIX] web_enterprise: reduce requests overhead when passing through SW
This commit reduces the HTTP requests' overhead when passing through the
fetch event in the ServiceWorker.

In a nutshell, before this commit every request is attempted by the SW
and only filtered out when the network request fails (to allow the
fallback to the "offline" page).

Now, it will first filter out the "navigation" request (cf. navigating
from one url to another one) and handle the "offline" case for them,
like before.

But all other requests will just pass through the SW without being
touched.

Doing so, it avoids the overhead (around 2-6ms) of being handled and
forwarded instead just passing through.

Note: the /web/offline is now self-contained and avoids requesting an
additional resource (aka. the logo).

closes odoo/enterprise#46876

X-original-commit: 50fb7ee4aa7937d2e31760e2a697bac314464f9d
Signed-off-by: Pierre Paridans (app) <app@odoo.com>