[FIX] account_asset: test indeterministic because wrong env
The test test_depreciation_schedule_hierarchy fail undeterministically.
This is due to the fact that a savepoint in _autopost_draft_entries
take an env at random.
This could cause the context to contain the value delay_account_group_sync
which would make the account not have groups.
We just remove the autopost, as it is useless (the moves are already posted)

Linked to runbot error 24474

closes odoo/enterprise#48350

X-original-commit: 4d0847abcc5cc380cc6900af395f5069854b4d12
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>