[FIX] mail_enterprise: restore cryptography 2.8 compatibility
This commit is a backport of https://github.com/odoo/enterprise/commit/191fd7ed65c0214dfe483dad8eb5be947961e0b9

A few internal systems are still running ubuntu 20.04 (focal) which
ships with cryptography 2.8. The backend parameter was made optional
(automatically falling back to the default backend) in 3.1.

Provide a backend explicitly so the module can be updated without
needing to upgrade the system (or install a non-distro package).

closes odoo/enterprise#46877

X-original-commit: db825a8a9596c421eccb859a777b72d9a1d3bb0a
Signed-off-by: Pierre Paridans (app) <app@odoo.com>