[FIX] account_reports: adapt the reports dispatching to section behavior
The introduction of sections in reports put in light an inappropriate behavior in the dispatching of the option buttons.

In a browser, when the currently displayed report is a section of another, the dispatch_report_action (controller.js in account_reports) would dynamically adapt the buttons to have them use the original report id. Hence, there would be no problem in the browser.

However, during testing, dispatch_report_action can't do its magic ; a button is called from the section instead of its parent, causing the test to fail.

Adapted the test to take sections into account.

closes odoo/enterprise#46846

X-original-commit: 8972edbdede1bfbfa39cbe85a0e49e385d357a37
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Camille Spiritus (casp) <casp@odoo.com>