[FIX] l10n_mx_edi_pos: fill the `CFDI Origin` upon POS refunds
From the POS, one can ask for a refund for an order. If this order was
invoiced with CFDI enabled, it has been sent to the government and has a
`Fiscal Folio` (an ID for this invoice with the government).

Mexico requires to fill the `Fiscal Folio` of the refunded invoice (as
the `CFDI Origin`) when creating a credit note.

task-3491567

closes odoo/enterprise#47027

X-original-commit: 6ba51ded388b478e27d1f31e9f13a3fb75d90c63
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>