[FIX] l10n_it_reports: fix Conto Economico
Some sections of the report were not computed correctly,
leading to inconsistent section totals, and ultimately,
an inconsistent report.

This was due to the fact that some account types were not
using the correct formula (sum vs -sum).
Moreover, when aggregating those lines, the aggregation formula
was sometimes incorrect as well.

This commit aims at fixing all these errors and harmonizing
the Conto Economico's structure with that of the other
localizations.

task-3492097
opw-3377392

closes odoo/enterprise#48217

X-original-commit: 7ebae31e5219f6e6a43ae5b26a61b64fd5242c03
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>