[FIX] l10n_tn_reports: Add omitted account 128 in Balance Sheet
Account 128 Accounting changes affecting retained earnings was
previously omitted from the balance sheet.

This commit adds it in 'Results carried forward'. (Since accounts 121
and 128 are the only ones starting with 12, we use the account code 12
to refer to them both.)

This is part of the wider task of re-balancing the balance sheets. I've
verified that with this change, the balance sheet is indeed balanced.

taskid:3239795

closes odoo/enterprise#48343

X-original-commit: e071eb095f8559d298cedd8ef0d5cf11a82d0b89
Signed-off-by: John Laterre (jol) <jol@odoo.com>