[FIX] sale_subscription: multi-currency issue in renewal
Before this commit, on a renewal quotation, changing the pricelist to a another one with different currency were not updating the order lines prices correctly. Now, when the pricelist is changed during a subscription renewal and the user clicks on the 'Update Prices' button, the prices are correctly recomputed and updated. This commit includes tests for ensuring this correct behavior.

Additionaly, when a rule defined for a pricelist was selected it wasn't being taken into account in priority because rules without pricelist were being selected to get the lowest price. Now, the selected pricelist and its time-based pricing rule is taken into priority in the new order line calculation. This solution part was obtained by backporting a fix (task-id 2939221) and its unit tests.

Also, before this commit, when the 'Recurrence' field of a subscription was updated, the 'Unit Price' field of subscription products order lines was not being recomputed and the same field for regular products (non recurring) order lines was being recomputed. After this commit, after changing the 'Recurrence field', the price is recalculated only for subscription products. Unit tests were added for ensuring this correct behavior.

Finally, the 'Unit Price' for product variants was not being calculated properly when the variants had a specific time-based pricing, for instance two variants V1 and V2 with different pricing and the 'Recurrence' field as 'Monthly', when added to the subscription, both of them had their 'Unit Price' field defined as the product default price. After this commit, the 'Unit Price' calculation for variants is working properly.

closes odoo/enterprise#39245

Task-id: 3253969
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>