[FIX] timesheet_grid: patch setTimeout to avoid undeterministic failing test
Before this commit, a test in `timesheet_grid` to check the grid view
used in `My Timesheets` fails before it does not find the overtime
displayed in the total column.

This commit patches the setTimeout function before the mouseover event
is debounced and so we have to wait the timeout is done before checking
if the overtime is displayed. With the patch, we no longer have to wait
in the test to have the overtime displayed when the mouseover is triggered
on total cell on the grid view.

runbot-22358

closes odoo/enterprise#47038

X-original-commit: d060b2e38cacfb2216249d3c71309ef402234d22
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>