[IMP] website_enterprise, *: keep user in backend on dark mode toggle
*web_enterprise

Since this task in community adds the user menu in the navbar, it is now
possible to toggle the dark mode from the website backend. However,
(1) toggling the dark mode reloads the page, and
(2) since [commit 1], nothing in the displayed URL differentiates the
frontend from the backend in the website app, meaning that reloading the
page from the backend actually leads to the frontend page.
This commit patches the `color_scheme_service` in order to remain in the
backend if it is toggled from the website app.

[commit 1]: https://github.com/odoo/odoo/commit/1c18b79972c3b0a97197b98390e0ba9fda703585

task-3381773

closes odoo/enterprise#46992

Related: odoo/odoo#133031
Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>