[IMP] sale_subscription: reduce query count
A query was added by the companion pr because a call to
`self.env.ref("sales_team.group_sale_salesman").id` was not in cache,
since get_description was not called anymore by `_track_get_fields`.
This was adding one query when calling _message_track.

This can easily be solved by making the same optimisation in
`_message_track` meaning that we have one less query for the exist.

closes odoo/enterprise#46997

Related: odoo/odoo#132950
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>