[IMP] web_studio: add computed options in Sidebar to adapt the display
This commit introduces the concept of 'Computed Options' in the sidebar
of the view editor. Those are options that are tied to another option.
Their value and visibility depends on another option present in the sidebar.

While computed options are used to adapt the display of the property in the
sidebar, the actual value is not changed in the view. This is only a way for
studio to improve the user experience by showing only relevant options, or
disabling those that are superseeded by others.

The super option mechanism can work with multiple layers of dependencies:

    - option A
    -    option C
    -       option D
    - option B

In this example, option D display and value will be computed from option C, which
is computed from the value of option A.

Tests have been added to verify the support of this feature for studio.

task-3469741

closes odoo/enterprise#47148

Related: odoo/odoo#134858
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>