[FIX] social_crm: hide "Create Lead" button for our own posts
Purpose
=======
The button "Create Lead" is supposed to be hidden for our own posts,
but it is always visible because we wrongly compute "isAuthor".

Technical
=========
A computed field is needed to fix the issue, because we don't have the
account media id in JS (or it will require a RPC call).

Task-3495203

closes odoo/enterprise#47204

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>