[ADD] appointment_google_calendar: generate Google url
Purpose
-------
Give the possibility to users to automatically generate
Google Meet urls for the events of an appointment type.
The feature already exists for Odoo Discuss links, extend the
behavior to add a Google Meet url support.

Specification
-------------
For the appointment to have a related Google Event with a
Google Meet url, the Odoo staff user will have to add a
Google Calendar ID and Secret in the Odoo settings and
synchronize its Google Account to the Odoo Calendar to be
fully synchronized.

Booking an appointment with the "Google Meet" option on its
appointment type will automatically create a related
Google Event and generate a google meet url.
The Odoo staff user who has synchronized its Google Calendar
account to Odoo will have the event generated both in Odoo
and in its Google Calendar.
The client who booked the appointment will receive an
invitation to the Google Event by email.

Removing the "Add to Google Agenda" button on the appointment
validation page when a Google Event has been generated to
prevent the client from having duplicated Google Events
(e.g. one created manually from the button and one
automatically added when the client is set as participant).

Task-32856011

Part-of: odoo/enterprise#46688