[FIX] knowledge: download demo survival guide file
The related commit made it possible to download a file from on /file
command of the knowledge app also via the mobile app. The implementation
was based on the assumption that is wasn't necessary to download a file
from outside the current database. That assumption is wrong regarding
the demo knowledge page where the /file demo file (Odoo Survival Guide)
is hosted on odoo.com (hence not the current database).

The problem is that the `download` function always send a POST XHR even
when there is no data to post (e.g. only download a file given its URL),
in case of the Odoo server, POST requests and more restrictive than GET
ones: they require a CSRF-token and perform some CORS verifications.
In the case of downloading via a POST to Odoo Survival Guide (hosted on
odoo.com) from another database, the CORS validation failed and the file
couldn't be downloaded.

Use `downloadFile` instead, it sends a `GET` XHR.

Note: The same problem was solved in saas-16.4 by embeding the Odoo
Survival Guide PDF inside the demo data, see second commit.

closes odoo/enterprise#47056

Reference-to: odoo/enterprise@b2588fa ([FIX] knowledge: allow file downloads in mobile app)
Reference-to: odoo/enterprise@ed4cc6735daa85e7d90a56971c119391f1e3df53 ([IMP] knowledge: improve onboarding)
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>