[IMP] l10n_mx,l10n_mx_edi: CFDI import & new `mx_tax_type` field on tax
Allow to import a CFDI xml from the dashboard or the invoices, bills,
refunds, credit notes view to generate the corresponding account_move.

Create a l10n_mx_edi.document with state 'sent' when importing a CFDI on
a sale journal, and 'received' on a purchase journal. This allows to
behave as the out_invoice/out_refund was created in Odoo and follow the
usual CFDI flow, but not for the in_invoice/in_refund which de not
require to be send to the government.

In addition, it allows to 'Request Cancel' for out_invoice/out_refund
(this button is not shown for in_invoice/in_refund since it can directly
be cancelled).

Finally, add a selection field `l10n_mx_tax_type` on account.tax. This
field is used in the CFDI attachment. This allows to avoid relying on
the name of the repartition line tags to export and import a CFDI.

task-3388347

closes odoo/enterprise#47321

Related: odoo/odoo#135215
Related: odoo/upgrade#5136
Signed-off-by: Laurent Smet (las) <las@odoo.com>