[IMP] mrp_workorder: remove session owner
Clean hook file to remove this references. The issue with it
is that we are not knowing which object will use the hook.
Also the function are not binded and this is not propagated
among the different functions.

The reload is not useful since the employees are in a state and
the interface refresh correctly

Add the functionality to remove the current owner of the session.
The purpose is to avoid someone impersonate the employee.

Task: 3456827 (nr 29)
Part-of: odoo/enterprise#48379