[IMP] sign: remove refusal allowed from sign
In sign, it is possible to refuse documents if the feature is enabled in the
module settings and the user marks the sign request as refusable when sending
the document from the wizard. This feature is quite hidden the way it is, so
it was decided to enable it for all documents.

task-3457176

Part-of: odoo/enterprise#46339