[IMP] sign: add one2many signers field
In sign, we were using a one2many to render all the signers and roles in the
wizard. However, this approach was prone to breaking every time the list
renderer was changed in web. We also had issues with autofocus and styles
breaking. This commit introduces a one2many field specific for sign that
removes the list rendered dependency, therefore giving more control to what
we can do and preventing the layout from breaking.

task-3457176

closes odoo/enterprise#46339

Related: odoo/upgrade#5079
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>