[FIX] web_studio: do preload field's value if invisible
Have some invisible fields in a form, list or kanban view
set as: `<field name="somefield" invisible="True" />`

In studio, go in the view tab in the sidebar, then check the show invisible box.

In the form view there was a crash because the invisible fields were not part of the new "unity"
read (`web_read`), such as the Many2one Field widget had the wrong data.

In the list and kanban view, there was not a crash, but the displayed data were wrong.

This commit adapts the handling of invisible fields in studio with the "unity" read. Specifically,
it considers all fields to be fetched, regardless of their visibility.

opw-3453701

closes odoo/enterprise#47129

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>