[FIX] web_studio: don't alter selection of a field when discarding
In studio, click on a selection field
Edit its possible values with the link in the sidebar

Add a value in the dialog's list, but then discard the changes.

Before this commit, the field in DB was not written, but in JS, when going
back to the selection values dialog the discarded value was still there.

This was because the dialog altered the field characteristics directly.

After this commit, the dialog operates on a copy of the selection values, and any discarded
value is not committed to the underlying object.

Part-of: odoo/enterprise#47137