[FIX] web_studio: concurrency: leaving studio while a rendering is ongoing
Operating in a FormView in Studio.

Leaving studio wipes the caches, in particular the viewCompiler's cache, that holds
an OWL template for each effective arch that we retrieve from get_views.

Now, click on the Close button to leave studio. Just *after* doing that,
trigger a rendering within studio (that can be very hardly done by a human though) *before* the new action, without
studio is ready.

Before this, commit, the studio formCompiler would have compiled the arch in place of the standard form compiler.
Thus, the standard formRenderer would receive the template prepared in the context of Studio, that is, with specific
components, crashing the whole thing.

After this commit, we reset the Compiler's cache when the new non-studio aciton is ready and studio has been left.
There is no crash any more

runbot-error-23046

closes odoo/enterprise#47137

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>