[IMP] l10n_pe_reports: Welcome RVIE and RCE
This pull request (PR) is centered around the newly enacted resolution Resolución 000190-2022/Sunat,
which brings modifications to the electronic sales and purchase reports.

These reports were initially introduced in a prior PR available at
https://github.com/odoo/enterprise/pull/37382, but they have now been superseded by this latest PR.

While the core content of these reports remains the same, there are notable alterations in the
downloadable TXT file. This includes the addition of new columns and a change in the column order.
All the required data already exists within Odoo; the key task here is to ensure its accurate
inclusion in the updated TXT file.

Here are the specific annexes for each report:

1. Report 14.4: You can access it here - https://drive.google.com/file/d/1HOAhlIx3ba_wtFvg9kufuCA6273KTVEy/view?usp=drive_link.

2. Report 8.4: The relevant document can be found at this link - https://drive.google.com/file/d/11WFdWHTY12DQKy8yxLzdwRl2zifTE8gD/view?usp=drive_link.

3. Report 8.5: You can view this report by following this link - https://drive.google.com/file/d/1_5dz44MUOryRgtAWTUWn3phmRTMexl4I/view?usp=drive_link.

closes odoo/enterprise#47195

Signed-off-by: Josse Colpaert <jco@odoo.com>