[IMP] l10n_us_hr_payroll: Add W2 form declaration
Purpose
=======

A W-2 is an official form that U.S. employers send to their employees and
the Internal Revenue Service (IRS) at the end of each year. It is one of
the most important reports involved in Payroll in the US. It details relevant
information about an employee's salary, such as how much money an employee
earned in that year, the total of different taxes retained, and benefits,
among others.

When tax season comes around, employees use the information on the W-2 to
help fill out their personal tax returns, ensuring they pay the right amount
of taxes or get a refund if they've overpaid. 

Specification
=============

Currently, in the US there are several online services that provide direct
filing of W2 and other reports directly with the government. So, as a
first iteration, we will offer the possibility to download a CSV file from
Odoo, containing specific information from Payroll needed to file the W2
online. (Example: Tax Bandits, Online File Taxes, Efile4Biz).

Taskid: 3434735
Part-of: odoo/enterprise#47223