[REF] account_reports: build column dict
Refactors '_build_column_dict' to take less arguments.

The idea is to have as little arguments as possible making the
function easy and straightforward to use on most cases.

The main difference is that instead of passing column values like
'figure_type' or 'blank_if_zero', we pass the whole column and then
recover the values we need inside the function.

closes odoo/enterprise#46494

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>