[FIX] point_of_sale: fix weighing scale
Before this commit, the weighing scale was not working properly. This
was caused by commit https://github.com/odoo-dev/odoo/commit/60f41b1f831248c58f554dbe40114ef0d4ce42c0, which modified the getAddProductOptions
function to return both the quantity and weight, instead of just the
quantity.

The "weight" value returned was not handled in the subsequent
"add_product" call stack, causing the weighing functionality to fail.

This commit fixes the issue by using the quantity value instead of
weight in the relevant code paths. In the point of sale context,
quantity and weight are interchangeable depending on the use case.

This commit fixes the associated tests to check the weight of added
order line.

opw-3487707

closes odoo/enterprise#47379

X-original-commit: 3a0a637faf9ec9cf73b8d8f15b001fbf3e1597f8
Related: odoo/odoo#135432
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>